added jenkinsfile

+++ bbtest/helpers/unit.py
@@ -19,7 +19,7 @@ class UnitHelper(object):
       "LOG_LEVEL": "DEBUG",
       "HTTP_PORT": "80",
       "POSTGRES_URL": "jdbc:postgresql://postgres:5432/openbank",
-      "PRIMARY_STORAGE_PATH": "{}/reports/blackbox-tests/meta".format(os.getcwd()),
+      "PRIMARY_STORAGE_PATH": "/data",
     }
 
   def get_arch(self):
@@ -39,10 +39,24 @@ class UnitHelper(object):
     self.docker = docker.from_env()
     self.context = context
 
-  def download(self):
-    failure = None
-    os.makedirs('/tmp/packages', exist_ok=True)
+  def install(self, file):
+    cwd = os.path.realpath('{}/../..'.format(os.path.dirname(__file__)))
 
+    (code, result, error) = execute(['dpkg', '-c', file])
+    if code != 0:
+      raise RuntimeError('code: {}, stdout: [{}], stderr: [{}]'.format(code, result, error))
+    else:
+      os.makedirs('{}/reports/blackbox-tests/meta'.format(cwd), exist_ok=True)
+      with open('{}/reports/blackbox-tests/meta/debian.data-warehouse.txt'.format(cwd), 'w') as fd:
+        fd.write(result)
+
+      result = [item for item in result.split(os.linesep)]
+      result = [item.rsplit('/', 1)[-1].strip() for item in result if "/lib/systemd/system/data-warehouse" in item]
+      result = [item for item in result if not item.endswith('unit.slice')]
+
+      self.units = result
+
+  def download(self):
     self.image_version = os.environ.get('IMAGE_VERSION', '')
     self.debian_version = os.environ.get('UNIT_VERSION', '')
 
@@ -52,16 +66,25 @@ class UnitHelper(object):
     assert self.image_version, 'IMAGE_VERSION not provided'
     assert self.debian_version, 'UNIT_VERSION not provided'
 
+    cwd = os.path.realpath('{}/../..'.format(os.path.dirname(__file__)))
+
+    self.binary = '{}/packaging/bin/data-warehouse_{}_{}.deb'.format(cwd, self.debian_version, self.arch)
+
+    if os.path.exists(self.binary):
+      self.install(self.binary)
+      return
+
+    os.makedirs(os.path.dirname(self.binary), exist_ok=True)
+
+    failure = None
     image = 'openbank/data-warehouse:{}'.format(self.image_version)
     package = '/opt/artifacts/data-warehouse_{}_{}.deb'.format(self.debian_version, self.arch)
-    target = '/tmp/packages/data-warehouse.deb'
-
     temp = tempfile.NamedTemporaryFile(delete=True)
     try:
       with open(temp.name, 'w') as fd:
         fd.write(str(os.linesep).join([
           'FROM alpine',
-          'COPY --from={} {} {}'.format(image, package, target)
+          'COPY --from={} {} {}'.format(image, package, self.binary)
         ]))
 
       image, stream = self.docker.images.build(fileobj=temp, rm=True, pull=False, tag='bbtest_artifacts-scratch')
@@ -78,26 +101,13 @@ class UnitHelper(object):
 
       tar_name = tempfile.NamedTemporaryFile(delete=True)
       with open(tar_name.name, 'wb') as fd:
-        bits, stat = scratch.get_archive(target)
+        bits, stat = scratch.get_archive(self.binary)
         for chunk in bits:
           fd.write(chunk)
 
       archive = tarfile.TarFile(tar_name.name)
-      archive.extract(os.path.basename(target), os.path.dirname(target))
-
-      (code, result, error) = execute(['dpkg', '-c', target])
-      if code != 0:
-        raise RuntimeError('code: {}, stdout: [{}], stderr: [{}]'.format(code, result, error))
-      else:
-        with open('reports/blackbox-tests/meta/debian.data-warehouse.txt', 'w') as fd:
-          fd.write(result)
-
-        result = [item for item in result.split(os.linesep)]
-        result = [item.rsplit('/', 1)[-1].strip() for item in result if "/lib/systemd/system/data-warehouse" in item]
-        result = [item for item in result if not item.endswith('unit.slice')]
-
-        self.units = result
-
+      archive.extract(os.path.basename(self.binary), os.path.dirname(self.binary))
+      self.install(self.binary)
       scratch.remove()
     except Exception as ex:
       failure = ex
@@ -122,16 +132,20 @@ class UnitHelper(object):
       fd.write(str(os.linesep).join("DATA_WAREHOUSE_{!s}={!s}".format(k, v) for (k, v) in options.items()))
 
   def collect_logs(self):
+    cwd = os.path.realpath('{}/../..'.format(os.path.dirname(__file__)))
+
+    os.makedirs('{}/reports/blackbox-tests/logs'.format(cwd), exist_ok=True)
+
     (code, result, error) = execute(['journalctl', '-o', 'cat', '--no-pager'])
     if code == 0:
-      with open('reports/blackbox-tests/logs/journal.log', 'w') as fd:
+      with open('{}/reports/blackbox-tests/logs/journal.log'.format(cwd), 'w') as fd:
         fd.write(result)
 
     for unit in set(self.__get_systemd_units() + self.units):
       (code, result, error) = execute(['journalctl', '-o', 'cat', '-u', unit, '--no-pager'])
       if code != 0 or not result:
         continue
-      with open('reports/blackbox-tests/logs/{}.log'.format(unit), 'w') as fd:
+      with open('{}/reports/blackbox-tests/logs/{}.log'.format(cwd, unit), 'w') as fd:
         fd.write(result)
 
   def teardown(self):
@@ -143,5 +157,5 @@ class UnitHelper(object):
   def __get_systemd_units(self):
     (code, result, error) = execute(['systemctl', 'list-units', '--no-legend'])
     result = [item.split(' ')[0].strip() for item in result.split(os.linesep)]
-    result = [item for item in result if "data-warehouse" in item]
+    result = [item for item in result if "data-warehouse" in item and not item.endswith('unit.slice')]
     return result
